% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_agr_value_pollination.R
\name{compute_agr_value_pollination}
\alias{compute_agr_value_pollination}
\title{COmputing unit values of agricultural production in Italian NUTS2}
\usage{
compute_agr_value_pollination(
  nuts = "Italia",
  ref_yr = 2019,
  last_yr = 2019,
  h = 3
)
}
\arguments{
\item{nuts}{A character vector. The names of the the Italian NUTS2 regions of interest. Names
must be in Italian and chosen among the folloqing list: `r knitr::combine_words(ecoservr::nuts2_codes$label)`.
`Bolzano` and `Trento` are treated as separate regions. Data on `Italia` as a whole can be retrieved
as well.}

\item{ref_yr}{A numeric value. The year at which price levels monetary values must be expressed.}

\item{last_yr}{An integer. The latest (meaning most recent) year from which computing average areas.}

\item{h}{An integer. A number specifying the number of years over which computing average areas. Default value
is `3`.}
}
\value{
A tibble with unit values, expressed at `ref_yr` prices, of the main crops
    that benefit from insect pollination for Italy and all Italian NUTS2 regions.
}
\description{
This function computes unit values of agricultural production to be used for the
calculation of the economic value of pollination services. Values are computed for crops that
benefit from pollination from insects. Crop categories are a special aggregation of crops
as defined by ISTAT and Eurostat, in order to reflect the classification used by Ispra to
quantify the ecosystems' contribution to crop pollination.
}
