% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_crop_shares.R
\name{compute_crop_shares}
\alias{compute_crop_shares}
\title{Computes the share of single crop classes on total agricultural area}
\usage{
compute_crop_shares(nuts = "Italia", last_yr = 2019, h = 3)
}
\arguments{
\item{nuts}{A character vector. The names of the the Italian NUTS2 regions of interest. Names
must be in Italian and chosen among the folloqing list: `r knitr::combine_words(ecoservr::nuts2_codes$label)`.
`Bolzano` and `Trento` are treated as separate regions. Data on `Italia` as a whole can be retrieved
as well.}

\item{last_yr}{An integer. The latest (meaning most recent) year from which computing average areas.}

\item{h}{An integer. A number specifying the number of years over which computing average areas. Default value
is `3`.}
}
\value{
A tibble with NUTS labels and codes and the share of insect pollinated crops on total UAA.
}
\description{
This function computes the ratio of the average areas of crops that benefit from insect pollination on total
utilized agricultural area in Italian NUTS2.
}
