% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_eco_con.R
\name{compute_eco_con}
\alias{compute_eco_con}
\title{Computing the ecosystems' contribution to food provisioning}
\usage{
compute_eco_con(nuts = "Italia", h = 3, last_yr, corine_code)
}
\arguments{
\item{nuts}{A character vector. The names of the the Italian NUTS2 regions of interest. Names
must be in Italian and chosen among the folloqing list: `r knitr::combine_words(nuts2_codes$label)`.
`Bolzano` and `Trento` are treated as separate regions. Data on `Italia` as a whole can be retrieved
as well.}

\item{h}{An integer. A number specifying the number of years over which computing average areas. Default value
is `3`.}

\item{last_yr}{An integer. The latest (meaning most recent) year from which computing average areas.}

\item{corine_code}{A character vector. Vector of the Corine classes (level 3) for which the average
area must be computed.}
}
\value{
a Tibble with Corine 3 codes and labels, total unit value of agricultural production (three-year average
expressed in euro/ha), the `eco_con_coeff` used and the contribution of ecosystems to food provision computed
as the product between `unit_value` and `eco_con_coeff`.
}
\description{
This function computes - using so-called eco-con-coefficients provided by the JRC of the European
Commission - the contribution of ecosystems in the provision of food, as a share of the total
value of agricultural production. Values are expressed in euro/ha (at the prices of `last_yr`) for each
of the Italian NUTS 2 regions and each of the Corine 3 classes selected.
}
\examples{

compute_eco_con(nuts = c("Umbria", "Puglia"), last_yr = 2019, corine_code = c("2.1.1", "2.2.2", "2.3.1"))
}
