% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_eco_con_forest.R
\name{compute_eco_con_forest}
\alias{compute_eco_con_forest}
\title{Computes the economic value of provisioning serivces from forest ecosystems}
\usage{
compute_eco_con_forest(
  nuts = "Italia",
  eco_con_timber = 0.97,
  nai_value = 0.402,
  ref_yr,
  forestry_data_yr = 2015
)
}
\arguments{
\item{nuts}{A character vector. The names of the the Italian NUTS2 regions of interest. Names
must be in Italian and chosen among the folloqing list: `r knitr::combine_words(nuts2_codes$label)`.
`Bolzano` and `Trento` are treated as separate regions. Data on `Italia` as a whole can be retrieved
as well.}

\item{eco_con_timber}{A numeric value. The ecosystems contribution to wood suppl. It can range between 0 (no ecosystem
contribution) and 1. The default value is `0.97`, the value calculated by the JRC see Vallecillo et al. (2018).}

\item{nai_value}{A numeric value. A factor (a number between 0 and 1) total forestry output must be multiplied by to
obtain the economic value of the net annual increment of wood in Italian forests.Default value is `0.402` fo
all Italian NUTS2 (no estimates at regional level are available).}

\item{ref_yr}{A numeric value. The year at which price levels monetary values must be expressed.}

\item{forestry_data_yr}{A numeric value. The year to which  latest data on Italian forests are available.
The default is `2015`, the latest available year. This parameter is important because, since the latest
available forest data are about that year, also economic values refer to the same year.}
}
\description{
This function computes the economic value of provisoning services (wood supply) provided by forest ecosystems
in Italian NUTS2 regions.
}
\seealso{
\url{https://www.inventarioforestale.org/it/}
}
