% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landscape_value.R
\name{landscape_value}
\alias{landscape_value}
\title{Compute landscape values for a specific corine land cover and maes ecosystem}
\usage{
landscape_value(
  nuts = "Italia",
  include_forest = FALSE,
  corine_code,
  policy_yr,
  ref_yr,
  epsilon = 0.2
)
}
\arguments{
\item{nuts}{A string vector. Italian NUTS to which the literature values must be transferred}

\item{include_forest}{A logical value. If `FALSE` (the default) the landscape value of urban areas is set to
zero. This is because, due to the trees, the view is obstructed.}

\item{corine_code}{A vector of numeric values. Corine classes for which landscape value must be computed}

\item{policy_yr}{A Numeric value. The year to which the value transfer must be performed to}

\item{ref_yr}{A Numeric value. The year at which levels prices must be expressed}

\item{epsilon}{A numeric value. The income elasticity of willingness to pay for environmental goods.}
}
\value{
A tibble contaning information on corine class, maes eocystem, original and transferred values and
    study and policy areas
}
\description{
This function computes region-specific landscape values startin from literature values for
corine land cover classes and MAES ecosystems for which literature values are available.
}
\examples{
landscape_value(nuts = c("Italia", "Umbria"),
corine_code = c(222, 321), policy_yr = 2019, ref_yr = 2019)
}
