% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maes_corine.R
\docType{data}
\name{maes_corine}
\alias{maes_corine}
\title{MAES - Corine correspondence table}
\format{
A dataframe with four variables

\describe{
  \item{corine3_code}{Corine class code}
  \item{corine3_label_en}{Corine class label in English}
  \item{corine3_label_it}{Corine class label in Italian}
  \item{maes}{Ecosystem label in English}
}
}
\usage{
maes_corine
}
\description{
Correspondence table between Corine classes (level 3) and the MAES ecosystem classification.
}
\keyword{datasets}
