% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/most_common_crop.R
\name{most_common_crop}
\alias{most_common_crop}
\title{Identifying the most common crop grown on arable land}
\usage{
most_common_crop(nuts = "Italia", h = 3, last_yr, lang = "it")
}
\arguments{
\item{nuts}{A character vector. The names of the the Italian NUTS2 regions of interest. Names
must be in Italian and chosen among the folloqing list: `r knitr::combine_words(ecoservr::nuts2_codes$label)`.
`Bolzano` and `Trento` are treated as separate regions. Data on `Italia` as a whole can be retrieved
as well.}

\item{h}{An integer. A number specifying the number of years over which computing average areas. Default value
is `3`.}

\item{last_yr}{An integer. The latest (meaning most recent) year from which computing average areas.}

\item{lang}{character string. Either `en` or `it`, to specify in which language Corine classes and crop labels must
be returned. Default value is `it`.}
}
\value{
A tibble with the average area occupied by crops grown on arable land.
}
\description{
This function identifies - on the basis of Eurostat official data - the most
common crop grown on arable land in Italian NUTS2 regions. The most common crop
is identified on the basis of the average crop areas of a user-defined time window.
}
\examples{

most_common_crop(last_yr = 2019)
}
