% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nuts2_codes.R
\docType{data}
\name{nuts2_codes}
\alias{nuts2_codes}
\title{Italian NUTS codes and labels}
\format{
## `nuts2_codes`
A data frame with 2 variables:
\describe{
  \item{code}{Eurostat code}
  \item{label}{NUTS2 label (in Italian)}
}
}
\usage{
nuts2_codes
}
\description{
A table containing the Eurostat codes of the Italian NUTS2 regions.
}
\keyword{datasets}
