% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pollination_value.R
\name{pollination_value}
\alias{pollination_value}
\title{Computes the economic value of pollination services for Italian NUTS2 regions}
\usage{
pollination_value(
  nuts = "Italia",
  last_yr = 2019,
  h = 3,
  ref_yr = 2019,
  pol_dependence = "ispra"
)
}
\arguments{
\item{nuts}{A character vector. The names of the the Italian NUTS2 regions of interest. Names
must be in Italian and chosen among the folloqing list: `r knitr::combine_words(nuts2_codes$label)`.
`Bolzano` and `Trento` are treated as separate regions. Data on `Italia` as a whole can be retrieved
as well.}

\item{last_yr}{An integer. The latest (meaning most recent) year from which computing average areas.}

\item{h}{An integer. A number specifying the number of years over which computing average areas. Default value
is `3`.}

\item{ref_yr}{A numeric value. The year at which price levels monetary values must be expressed.}

\item{pol_dependence}{A string. It allows the user to choose between crop-specific coefficients
provided by either the JRS (`jrc`) or ISPRA (`ispra`) or a nation-wide average (still based on
Ispra data: `ita_avg` of 10.4\%)}
}
\value{
A tibble with the economic value of pollination services in each of the NUTS2 selected.
    Values are expressed in euro/ha at constant prices (as specified through `ref_yr`)
}
\description{
This function computes the economic value of the contribution of the pollination service to
agricultural production. Starting from the average value of the insect pollinated crops in a
given time horizon (user-defined) it computes the average value of the ecosystem service
"pollination" on the basis of the share of each insect pollinated crop on total regional
utilized agricultural area and crop-specific pollination contribution coeffcients (how much
of a crop's total output is due to pollinaton services) taking into account the fact
that not all of pollination demand is actually met (see table 4.5 in Vallecillo, S et al. 2018)
}
\examples{
pollination_value(nuts = c("Italia", "Umbria"))
}
