% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provisioning_value.R
\name{provisioning_value}
\alias{provisioning_value}
\title{Computing the ecosystems' contribution to the provision of food and wood}
\usage{
provisioning_value(
  nuts = "Italia",
  h = 3,
  last_yr,
  ref_yr = 2019,
  corine_code,
  lang = "it"
)
}
\arguments{
\item{nuts}{A character vector. The names of the the Italian NUTS2 regions of interest. Names
must be in Italian and chosen among the folloqing list: `r knitr::combine_words(ecoservr::nuts2_codes$label)`.
`Bolzano` and `Trento` are treated as separate regions. Data on `Italia` as a whole can be retrieved
as well.}

\item{h}{An integer. A number specifying the number of years over which computing average areas. Default value
is `3`.}

\item{last_yr}{An integer. The latest (meaning most recent) year from which computing average areas.}

\item{ref_yr}{A numeric value. The year at which price levels monetary values must be expressed.}

\item{corine_code}{A numeric vector. Vector of the Corine classes (level 3) for which the average
area must be computed.}

\item{lang}{character string. Either `en` or `it`, to specify in which language Corine classes and crop labels must
be returned. Default value is `it`.}
}
\value{
a Tibble with Corine 3 codes and labels, total unit value of agricultural production (three-year average
expressed in euro/ha), the `eco_con_coeff` used and the contribution of ecosystems to food provision computed
as the product between `unit_value` and `eco_con_coeff`.
}
\description{
This function computes - using so-called eco-con-coefficients provided by the JRC of the European
Commission - the contribution of ecosystems in the provision of food and wood. Values are expressed in euro/ha (at the prices of `ref_yr`)
for each of the Italian NUTS 2 regions and each of the Corine 3 classes selected.
}
\examples{

provisioning_value(nuts = c("Umbria", "Puglia"),
last_yr = 2019, ref_yr = 2019, corine_code = c(211, 222, 231))
}
